refactor(editor): restructure editor component and add new features
Restructure the editor component to improve maintainability and add new features such as slash commands and a floating bubble menu.
