+++ apps/app/package.json
@@ -22,6 +22,11 @@
     "@tanstack/react-query": "^5.75.7",
     "@tanstack/react-query-devtools": "^5.75.7",
     "@tanstack/react-table": "^8.21.3",
+    "@tiptap/extension-highlight": "^2.12.0",
+    "@tiptap/extension-typography": "^2.12.0",
+    "@tiptap/pm": "^2.12.0",
+    "@tiptap/react": "^2.12.0",
+    "@tiptap/starter-kit": "^2.12.0",
     "@trpc/client": "^11.1.2",
     "@trpc/server": "^11.1.2",
     "@trpc/tanstack-react-query": "^11.1.2",
