+++ apps/app/src/app/editor/page.tsx
@@ -1,10 +1,10 @@
-import Editor from "@/components/editor/editor";
+import Editor from "@/components/editor";
 
 export default function EditorPage() {
   return (
     <div>
       <h1>Editor Page</h1>
-      <Editor placeholder="hello" />
+      <Editor placeholder="Add description..." />
     </div>
   );
 }
