+++ apps/app/src/components/editor/components/bubble-menu/node-selector.tsx
@@ -112,7 +112,7 @@ export const NodeSelector = ({ open, onOpenChange }: NodeSelectorProps) => {
         asChild
         className="gap-2 rounded-none border-none hover:bg-accent focus:ring-0"
       >
-        <Button variant="ghost" className="gap-2">
+        <Button variant="ghost" className="gap-2 ">
           <span className="whitespace-nowrap text-sm">{activeItem.name}</span>
           <ChevronDown className="h-4 w-4" />
         </Button>
