+++ apps/app/src/components/editor/components/slash-commands/slash-command.tsx
