+++ apps/app/src/components/editor/extensions/extentions.ts
@@ -1,14 +1,13 @@
+import Typography from "@tiptap/extension-typography";
 import {
   AIHighlight,
   CharacterCount,
   CodeBlockLowlight,
-  Color,
   CustomKeymap,
   GlobalDragHandle,
   HighlightExtension,
   HorizontalRule,
   Mathematics,
-  Placeholder,
   StarterKit,
   TaskItem,
   TaskList,
@@ -27,8 +26,6 @@ import { common, createLowlight } from "lowlight";
 
 //TODO I am using cx here to get tailwind autocomplete working, idk if someone else can write a regex to just capture the class key in objects
 const aiHighlight = AIHighlight;
-//You can overwrite the placeholder with your own configuration
-const placeholder = Placeholder;
 const tiptapLink = TiptapLink.configure({
   HTMLAttributes: {
     class: cx(
@@ -152,7 +149,6 @@ const characterCount = CharacterCount.configure();
 
 export const defaultExtensions = [
   starterKit,
-  placeholder,
   tiptapLink,
   tiptapImage,
   updatedImage,
@@ -168,7 +164,7 @@ export const defaultExtensions = [
   TiptapUnderline,
   HighlightExtension,
   TextStyle,
-  Color,
   CustomKeymap,
   GlobalDragHandle,
+  Typography,
 ];
